Moved logging components into separate folder to clean up structure.
